Source/requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.18
cmsplugin-filer==1.1.3
dj-database-url==0.4.1
Django==1.8.14
django-appconf==1.0.2
django-bootstrap3-datetimepicker==2.2.3
django-celery==3.1.17
django-celery-email==1.1.5
django-classy-tags==0.8.0
django-cms==3.4.0
django-datetime-widget==0.9
django-disqus==0.5
django-endless-pagination==2.0
django-filer==1.2.5
django-formtools==1.0
django-mail-queue==2.2.2
django-mptt==0.8.6
django-polymorphic==0.8.1
django-post-office==2.0.8
django-redis==4.5.0
django-sekizai==0.10.0
Django-Select2==4.3.2
django-social-widgets==0.4.0
django-tinymce==2.4.0
django-treebeard==4.0.1
django-twitter-tag==1.2.1
django-uwsgi==0.1.4
djangocms-admin-style==1.2.4
djangocms-attributes-field==0.1.1
djangocms-column==1.7.0
djangocms-googlemap==0.5.2
djangocms-installer==0.9
djangocms-link==2.0.0
djangocms-snippet==1.9.1
djangocms-style==1.7.0
djangocms-text-ckeditor==3.2.1
djangocms-video==2.0.1
easy-thumbnails==2.3
feedparser==5.2.1
funcsigs==1.0.2
html5lib==0.9999999
jsonfield==1.0.3
kombu==3.0.37
mock==2.0.0
oauthlib==2.0.1
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.6.2
pytz==2016.6.1
redis==2.10.5
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
twitter==1.17.1
twython==3.4.0
tzlocal==1.2.2
Unidecode==0.4.19
uwsgidecorators==1.1.0
